Moved path detection functions into utils.py and loaded them for steps.py and event_proxy.py. This allows for dynamic detection of test scripts and detection of where the images are.
Gui was changed so it can receive the unique test specification for the buttons in the form of the options tab.
The input of the options tab now determines the text on the buttons.
Utils get_project_root() was updated to be more resillient to different directory structure instead of being a hardcoded three directories back.