Made the global and local now work with no errors. Here it is that global values are put into the globals one and not actually in the parameter one. It may seem obvious to put it in parameter, but this one doesnt do much. However by the construction of the yaml reader, this line is still required to avoid a breakdown, even if empty.
Additionally flexible buttons have been incorporated using @msuminsk's suggestion so that more than two are possible to create, and if not given a name label, then their value is set to be the capitalized version of the label. The output of the function however is still inflexible and any one click on button is considered a response and then it goes to the next step in the sequence.