Applied both a start and stop button along with opening o files. all should be disconnected from runtime. GUI sends commands to the queue, where a thread reads specific commands and handles creation of new threads and the turn off of them.
Starting the button recreates a runtime thread, however as of the current version, it runs purely on a recipe thread which is instantiated at startup time. This makes it run from the current step of the recipe rather than a full recipe reset, but also means that other recipies are written over after loaded.